+++ src/form-elements/FormElementCamera.tsx
@@ -25,6 +25,9 @@ import ProgressBar from '../components/renderer/attachments/ProgressBar'
 import useElementAriaDescribedby from '../hooks/useElementAriaDescribedby'
 import MaterialIcon from '../components/MaterialIcon'
 import FormElementValidationMessage from '../components/renderer/FormElementValidationMessage'
+import CropModal from '../components/ImageCropper/CropModal'
+import { Area } from 'react-easy-crop'
+import { generateCroppedImageBlob } from '../components/ImageCropper'
 
 type Props = {
   id: string
@@ -54,6 +57,8 @@ function FormElementCamera({
   })
   const [isAnnotating, setIsAnnotating, clearIsAnnotating] =
     useBooleanState(false)
+  const [isCropping, setIsCropping, clearIsCropping] = useBooleanState(false)
+
   const fileInputRef = React.useRef<HTMLInputElement>(null)
 
   const clearImage = React.useCallback(() => {
@@ -181,6 +186,7 @@ function FormElementCamera({
     loadAttachmentUrlError,
     canDownload,
     progress,
+    contentType,
   } = useAttachment(
     value,
     element,
@@ -285,6 +291,44 @@ function FormElementCamera({
     [attachmentUrl, clearIsAnnotating, element, onChange],
   )
 
+  const handleSaveCrop = React.useCallback(
+    async (croppedAreaPixels: Area) => {
+      clearIsCropping()
+      if (!attachmentUrl) return
+
+      setState({
+        isLoading: true,
+      })
+
+      try {
+        const croppedImage = await generateCroppedImageBlob({
+          croppedAreaPixels,
+          imgSrc: attachmentUrl,
+          fileType: contentType,
+        })
+
+        if (!croppedImage) throw new Error('Cropped image is null')
+        const attachment = prepareNewAttachment(
+          croppedImage,
+          'photo.png',
+          element,
+        )
+        onChange(element, {
+          value: attachment,
+        })
+        setState({
+          isLoading: false,
+        })
+      } catch (error) {
+        setState({
+          cameraError: error as Error,
+          isLoading: false,
+        })
+      }
+    },
+    [attachmentUrl, clearIsCropping, contentType, element, onChange],
+  )
+
   const progressTooltipRef = React.useRef<HTMLDivElement>(null)
   return (
     <>
@@ -307,6 +351,7 @@ function FormElementCamera({
                   isLoading={isLoading}
                   element={element}
                   onAnnotate={setIsAnnotating}
+                  onCrop={setIsCropping}
                   canDownload={canDownload}
                   progress={progress}
                 />
@@ -388,6 +433,14 @@ function FormElementCamera({
         />
       )}
 
+      {isCropping && attachmentUrl && (
+        <CropModal
+          imageSrc={attachmentUrl}
+          onClose={clearIsCropping}
+          onSave={handleSaveCrop}
+        />
+      )}
+
       {cameraError && (
         <Modal
           isOpen
@@ -431,11 +484,13 @@ const DisplayImage = React.memo(function DisplayImage({
   isLoading,
   element,
   onAnnotate,
+  onCrop,
   progress,
-}: ReturnType<typeof useAttachment> & {
+}: Omit<ReturnType<typeof useAttachment>, 'contentType'> & {
   element: FormTypes.CameraElement
   isLoading: boolean
   onAnnotate: () => void
+  onCrop: () => void
   progress: number | undefined
 }) {
   if (uploadErrorMessage) {
@@ -485,16 +540,28 @@ const DisplayImage = React.memo(function DisplayImage({
           crossOrigin="anonymous"
           alt={`${element.label}: Attachment`}
         />
-        <button
-          type="button"
-          className="button is-primary ob-camera__annotate-button cypress-annotate-button"
-          onClick={onAnnotate}
-          disabled={element.readOnly}
-        >
-          <span className="icon">
-            <MaterialIcon>brush</MaterialIcon>
-          </span>
-        </button>
+        <div className="ob-image-file__actions">
+          <button
+            type="button"
+            className="button is-primary ob-camera__crop-button cypress-crop-button"
+            onClick={onCrop}
+            disabled={element.readOnly}
+          >
+            <span className="icon">
+              <MaterialIcon>crop</MaterialIcon>
+            </span>
+          </button>
+          <button
+            type="button"
+            className="button is-primary ob-camera__annotate-button cypress-annotate-button"
+            onClick={onAnnotate}
+            disabled={element.readOnly}
+          >
+            <span className="icon">
+              <MaterialIcon>brush</MaterialIcon>
+            </span>
+          </button>
+        </div>
       </>
     )
   }